api/pyproject.toml
[project]
name = "aviary-api"
version = "0.2.0"
description = "Aviary API — user-facing REST/WebSocket gateway (Phase 2)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.6.0",
    "websockets>=13.0",
    "aviary-shared",
    "httpx>=0.28.0",
    "redis>=5.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
